lucetc: move new stuff to root. rustfmt.
