--- lucetc/src/new/mod.rs
@@ -1,17 +1,0 @@
-mod compiler;
-mod decls;
-mod function;
-mod heap;
-mod module;
-mod name;
-mod output;
-mod pointer;
-mod runtime;
-mod sparsedata;
-mod stack_probe;
-mod table;
-mod traps;
-
-pub use compiler::{Compiler, OptLevel};
-pub use output::{CraneliftFuncs, ObjectFile};
-pub use heap::HeapSettings;
