+++ lucetc/src/pointer.rs
@@ -1,4 +1,3 @@
-
 use cranelift_codegen::ir;
 
 pub const POINTER_SIZE: usize = 8;
