+++ lucetc/src/traps.rs
